[#262] Modify install.sh to check that octo.conf doesn't exist and only copy it to $ydb_dist/plugin/etc if true. This prevents user's configuration from getting overwritten on every new version of octo.
Changing the CMake scripts so that `make install` won't overwrite
octo.conf doesn't seem to be possible, so that isn't addressed in this
change.

Update README.md to include instructions for the binary install
and running install.sh. Also, fix some Markdown formatting.

+++ src/aux/install.sh
@@ -54,7 +54,11 @@ if [[ $install =~ ^[Yy]$ ]]; then
         install -d $ydb_dist/plugin/o/utf8
         install -D -m 644 o/* $ydb_dist/plugin/o/utf8
         install -d $ydb_dist/plugin/etc
-        install -D -m 644 etc/* $ydb_dist/plugin/etc
+        if [ -f "$ydb_dist/plugin/etc/octo.conf" ]; then
+                install -D -m 644 $(ls -d etc/* | grep -v octo.conf) $ydb_dist/plugin/etc
+        else
+                install -D -m 644 etc/* $ydb_dist/plugin/etc
+        fi
 fi
 
 # Import Postgres metatadata to current global directory
